agent-ds/requirements.txt
# Agent DS - Autonomous Red-Team CLI AI Framework
# Core Dependencies

# CLI Framework
click==8.1.7
colorama==0.4.6
rich==13.7.0
typer==0.9.0

# AI/ML Libraries
torch==2.0.1
transformers==4.35.2
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.5
tensorflow==2.13.0
keras==2.13.1
joblib==1.3.2
pandas==2.1.4

# Network & Security Tools
python-nmap==0.7.1
python-masscan==0.1.6
requests==2.31.0
beautifulsoup4==4.12.2
pycvesearch==1.2
shodan==1.30.1

# Database & Storage
sqlite3
sqlalchemy==2.0.23
pymongo==4.6.0

# Web & API Tools
flask==3.0.0
aiohttp==3.9.1
selenium==4.15.2
playwright==1.40.0

# Exploitation & Testing
pwntools==4.11.0
paramiko==3.4.0
python-libnmap==0.7.3

# Reporting & Documentation
jinja2==3.1.2
reportlab==4.0.7
weasyprint==60.2
matplotlib==3.7.4
plotly==5.17.0

# Security & Cryptography
cryptography==41.0.7
pycryptodome==3.19.0
passlib==1.7.4
jwt==1.3.1
PyJWT==2.8.0

# Utility Libraries
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2
psutil==5.9.6
asyncio
concurrent.futures

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1